refactor(structure): reorganize project into feature-based architecture
