+++ src/components/query/ResultDisplay.jsx
